feat: add utility functions for responsive layout resolution and section heading injection
